Split verifyer away from deployer
