+++ raiden_contracts/deploy/__main__.py
@@ -5,13 +5,11 @@ import functools
 import json
 import logging
 from logging import getLogger
-from pathlib import Path
-from typing import Any, Dict, List, Optional
+from typing import Dict, List, Optional
 
 import click
 from click import BadParameter
 from eth_utils import denoms, encode_hex, is_address, to_checksum_address
-from mypy_extensions import TypedDict
 from web3 import HTTPProvider, Web3
 from web3.contract import Contract
 from web3.middleware import geth_poa_middleware
@@ -31,16 +29,13 @@ from raiden_contracts.constants import (
 from raiden_contracts.contract_manager import (
     ContractManager,
     contract_version_string,
-    contracts_deployed_path,
     contracts_precompiled_path,
-    get_contracts_deployed,
 )
 from raiden_contracts.deploy.contract_deployer import ContractDeployer
-from raiden_contracts.utils.bytecode import runtime_hexcode
+from raiden_contracts.deploy.contract_verifyer import ContractVerifyer, DeployedContracts
 from raiden_contracts.utils.private_key import get_private_key
 from raiden_contracts.utils.signature import private_key_to_address
 from raiden_contracts.utils.transaction import check_successful_tx
-from raiden_contracts.utils.type_aliases import Address
 
 LOG = getLogger(__name__)
 
@@ -145,66 +140,6 @@ def main():
     pass
 
 
-def store_and_verify_deployment_info_raiden(
-        deployed_contracts_info: 'DeployedContracts',
-        deployer: ContractDeployer,
-        contracts_version: Optional[str],
-        save_info: bool,
-):
-    if save_info is True:
-        store_deployment_info(
-            deployment_file_path=contracts_deployed_path(
-                chain_id=int(deployer.web3.version.network),
-                version=contracts_version,
-                services=False,
-            ),
-            deployment_info=deployed_contracts_info,
-        )
-        verify_deployed_contracts_in_filesystem(
-            deployer.web3,
-            deployer.contract_manager,
-        )
-    else:
-        verify_deployment_data(
-            web3=deployer.web3,
-            contract_manager=deployer.contract_manager,
-            deployment_data=deployed_contracts_info,
-        )
-
-
-def store_and_verify_deployment_info_services(
-        contracts_version: Optional[str],
-        deployer: ContractDeployer,
-        deployed_contracts_info: 'DeployedContracts',
-        save_info: bool,
-        token_address: str,
-        user_deposit_whole_limit: int,
-):
-    if save_info is True:
-        store_deployment_info(
-            deployment_file_path=contracts_deployed_path(
-                chain_id=int(deployer.web3.version.network),
-                version=contracts_version,
-                services=True,
-            ),
-            deployment_info=deployed_contracts_info,
-        )
-        verify_deployed_service_contracts_in_filesystem(
-            web3=deployer.web3,
-            contract_manager=deployer.contract_manager,
-            token_address=token_address,
-            user_deposit_whole_balance_limit=user_deposit_whole_limit,
-        )
-    else:
-        verify_service_contracts_deployment_data(
-            web3=deployer.web3,
-            contract_manager=deployer.contract_manager,
-            token_address=token_address,
-            user_deposit_whole_balance_limit=user_deposit_whole_limit,
-            deployment_data=deployed_contracts_info,
-        )
-
-
 def contract_version_with_max_token_networks(version: Optional[str]) -> bool:
     manager = ContractManager(contracts_precompiled_path(version))
     abi = manager.get_contract_abi(CONTRACT_TOKEN_NETWORK_REGISTRY)
@@ -291,8 +226,7 @@ def raiden(
         for contract_name, info in deployed_contracts_info['contracts'].items()
     }
 
-    store_and_verify_deployment_info_raiden(
-        deployer=deployer,
+    deployer.store_and_verify_deployment_info_raiden(
         deployed_contracts_info=deployed_contracts_info,
         contracts_version=contracts_version,
         save_info=save_info,
@@ -355,7 +289,7 @@ def services(
         for contract_name, info in deployed_contracts_info['contracts'].items()
     }
 
-    store_and_verify_deployment_info_services(
+    deployer.store_and_verify_deployment_info_services(
         contracts_version=contracts_version,
         deployer=deployer,
         deployed_contracts_info=deployed_contracts_info,
@@ -526,11 +460,8 @@ def verify(ctx, rpc_provider, contracts_version):
     web3.middleware_stack.inject(geth_poa_middleware, layer=0)
     print('Web3 provider is', web3.providers[0])
 
-    contract_manager = ContractManager(contracts_precompiled_path(contracts_version))
-    verify_deployed_contracts_in_filesystem(
-        web3=web3,
-        contract_manager=contract_manager,
-    )
+    verifyer = ContractVerifyer(web3=web3, contracts_version=contracts_version)
+    verifyer.verify_deployed_contracts_in_filesystem()
 
 
 def deployed_data_from_receipt(receipt, constructor_arguments):
@@ -728,268 +659,6 @@ def register_token_network(
         ),
     )
     return token_network_address
-
-
-def store_deployment_info(deployment_file_path: Path, deployment_info: 'DeployedContracts'):
-    with deployment_file_path.open(mode='w') as target_file:
-        target_file.write(json.dumps(deployment_info))
-
-    print(
-        f'Deployment information for chain id = {deployment_info["chain_id"]} '
-        f' has been updated at {deployment_file_path}.',
-    )
-
-
-def verify_deployment_data(
-        web3: Web3,
-        contract_manager: ContractManager,
-        deployment_data,
-):
-    chain_id = int(web3.version.network)
-    assert deployment_data is not None
-
-    assert contract_manager.version_string() == deployment_data['contracts_version']
-    assert chain_id == deployment_data['chain_id']
-
-    verify_deployed_contract(
-        web3=web3,
-        contract_manager=contract_manager,
-        deployment_data=deployment_data,
-        contract_name=CONTRACT_ENDPOINT_REGISTRY,
-    )
-
-    secret_registry, _ = verify_deployed_contract(
-        web3=web3,
-        contract_manager=contract_manager,
-        deployment_data=deployment_data,
-        contract_name=CONTRACT_SECRET_REGISTRY,
-    )
-
-    token_network_registry, constructor_arguments = verify_deployed_contract(
-        web3=web3,
-        contract_manager=contract_manager,
-        deployment_data=deployment_data,
-        contract_name=CONTRACT_TOKEN_NETWORK_REGISTRY,
-    )
-
-    # We need to also check the constructor parameters against the chain
-    assert to_checksum_address(
-        token_network_registry.functions.secret_registry_address().call(),
-    ) == secret_registry.address
-    assert secret_registry.address == constructor_arguments[0]
-    assert token_network_registry.functions.chain_id().call() == constructor_arguments[1]
-    assert token_network_registry.functions.settlement_timeout_min().call() == \
-        constructor_arguments[2]
-    assert token_network_registry.functions.settlement_timeout_max().call() == \
-        constructor_arguments[3]
-
-    return True
-
-
-def verify_deployed_contracts_in_filesystem(
-        web3: Web3,
-        contract_manager: ContractManager,
-):
-    chain_id = int(web3.version.network)
-
-    deployment_data = get_contracts_deployed(
-        chain_id=chain_id,
-        version=contract_manager.contracts_version,
-    )
-    deployment_file_path = contracts_deployed_path(
-        chain_id=chain_id,
-        version=contract_manager.contracts_version,
-    )
-    assert deployment_data is not None
-
-    if verify_deployment_data(web3, contract_manager, deployment_data):
-        print(f'Deployment info from {deployment_file_path} has been verified and it is CORRECT.')
-
-
-def verify_service_contracts_deployment_data(
-        web3: Web3,
-        contract_manager: ContractManager,
-        token_address: str,
-        user_deposit_whole_balance_limit: int,
-        deployment_data: 'DeployedContracts',
-):
-    chain_id = int(web3.version.network)
-    assert deployment_data is not None
-
-    assert contract_manager.version_string() == deployment_data['contracts_version']
-    assert chain_id == deployment_data['chain_id']
-
-    service_bundle, constructor_arguments = verify_deployed_contract(
-        web3=web3,
-        contract_manager=contract_manager,
-        deployment_data=deployment_data,
-        contract_name=CONTRACT_SERVICE_REGISTRY,
-    )
-    assert to_checksum_address(service_bundle.functions.token().call()) == token_address
-    assert token_address == constructor_arguments[0]
-
-    user_deposit, constructor_arguments = verify_deployed_contract(
-        web3=web3,
-        contract_manager=contract_manager,
-        deployment_data=deployment_data,
-        contract_name=CONTRACT_USER_DEPOSIT,
-    )
-    assert len(constructor_arguments) == 2
-    assert to_checksum_address(user_deposit.functions.token().call()) == token_address
-    assert token_address == constructor_arguments[0]
-    assert user_deposit.functions.whole_balance_limit().call() == user_deposit_whole_balance_limit
-    assert user_deposit_whole_balance_limit == constructor_arguments[1]
-
-    monitoring_service, constructor_arguments = verify_deployed_contract(
-        web3,
-        contract_manager,
-        deployment_data,
-        CONTRACT_MONITORING_SERVICE,
-    )
-    assert len(constructor_arguments) == 3
-    assert to_checksum_address(monitoring_service.functions.token().call()) == token_address
-    assert token_address == constructor_arguments[0]
-
-    assert to_checksum_address(
-        monitoring_service.functions.service_registry().call(),
-    ) == service_bundle.address
-    assert service_bundle.address == constructor_arguments[1]
-
-    assert to_checksum_address(
-        monitoring_service.functions.user_deposit().call(),
-    ) == user_deposit.address
-    assert user_deposit.address == constructor_arguments[2]
-
-    one_to_n, constructor_arguments = verify_deployed_contract(
-        web3=web3,
-        contract_manager=contract_manager,
-        deployment_data=deployment_data,
-        contract_name=CONTRACT_ONE_TO_N,
-    )
-    assert to_checksum_address(
-        one_to_n.functions.deposit_contract().call(),
-    ) == user_deposit.address
-    assert user_deposit.address == constructor_arguments[0]
-    assert len(constructor_arguments) == 1
-
-    # Check that UserDeposit.init() had the right effect
-    onchain_msc_address = to_checksum_address(user_deposit.functions.msc_address().call())
-    assert onchain_msc_address == monitoring_service.address, \
-        f'MSC address found onchain: {onchain_msc_address}, expected: {monitoring_service.address}'
-    assert to_checksum_address(
-        user_deposit.functions.one_to_n_address().call(),
-    ) == one_to_n.address
-
-    return True
-
-
-def verify_deployed_service_contracts_in_filesystem(
-        web3: Web3,
-        contract_manager: ContractManager,
-        token_address: str,
-        user_deposit_whole_balance_limit: int,
-):
-    chain_id = int(web3.version.network)
-
-    deployment_data = get_contracts_deployed(
-        chain_id=chain_id,
-        version=contract_manager.contracts_version,
-        services=True,
-    )
-    deployment_file_path = contracts_deployed_path(
-        chain_id=chain_id,
-        version=contract_manager.contracts_version,
-        services=True,
-    )
-    assert deployment_data is not None
-
-    if verify_service_contracts_deployment_data(
-            web3=web3,
-            contract_manager=contract_manager,
-            token_address=token_address,
-            user_deposit_whole_balance_limit=user_deposit_whole_balance_limit,
-            deployment_data=deployment_data,
-    ):
-        print(f'Deployment info from {deployment_file_path} has been verified and it is CORRECT.')
-
-
-def verify_deployed_contract(
-        web3: Web3,
-        contract_manager: ContractManager,
-        deployment_data: 'DeployedContracts',
-        contract_name: str,
-) -> Contract:
-    """ Verify deployment info against the chain
-
-    Verifies:
-    - the runtime bytecode - precompiled data against the chain
-    - information stored in deployment_*.json against the chain,
-    except for the constructor arguments, which have to be checked
-    separately.
-
-    Returns: (onchain_instance, constructor_arguments)
-    """
-    contracts = deployment_data['contracts']
-
-    contract_address = contracts[contract_name]['address']
-    contract_instance = web3.eth.contract(
-        abi=contract_manager.get_contract_abi(contract_name),
-        address=contract_address,
-    )
-
-    # Check that the deployed bytecode matches the precompiled data
-    blockchain_bytecode = web3.eth.getCode(contract_address).hex()
-    compiled_bytecode = runtime_hexcode(
-        contracts_manager=contract_manager,
-        name=contract_name,
-    )
-    assert blockchain_bytecode == compiled_bytecode
-
-    print(
-        f'{contract_name} at {contract_address} '
-        f'matches the compiled data from contracts.json',
-    )
-
-    # Check blockchain transaction hash & block information
-    receipt = web3.eth.getTransactionReceipt(
-        contracts[contract_name]['transaction_hash'],
-    )
-    assert receipt['blockNumber'] == contracts[contract_name]['block_number'], (
-        f'We have block_number {contracts[contract_name]["block_number"]} '
-        f'instead of {receipt["blockNumber"]}'
-    )
-    assert receipt['gasUsed'] == contracts[contract_name]['gas_cost'], (
-        f'We have gasUsed {contracts[contract_name]["gas_cost"]} '
-        f'instead of {receipt["gasUsed"]}'
-    )
-    assert receipt['contractAddress'] == contracts[contract_name]['address'], (
-        f'We have contractAddress {contracts[contract_name]["address"]} '
-        f'instead of {receipt["contractAddress"]}'
-    )
-
-    # Check the contract version
-    version = contract_instance.functions.contract_version().call()
-    assert version == deployment_data['contracts_version'], \
-        f'got {version} expected {deployment_data["contracts_version"]}.' \
-        f'contract_manager has contracts_version {contract_manager.contracts_version}'
-
-    return contract_instance, contracts[contract_name]['constructor_arguments']
-
-
-# Classes for static type checking of deployed_contracts dictionary.
-
-class DeployedContract(TypedDict):
-    address: Address
-    transaction_hash: str
-    block_number: int
-    gas_cost: int
-    constructor_arguments: Any
-
-
-class DeployedContracts(TypedDict):
-    chain_id: int
-    contracts: Dict[str, DeployedContract]
-    contracts_version: str
 
 
 if __name__ == '__main__':
